Allocate underlying pool_gang only when actually needed.
A pool_gang uses a file descriptor (to implement wakeup).  Since file
descriptors are a somewhat scarce resource (more so than memory), we
don't want them allocated if they don't need to be.  In particular, if
clearGang is called, we don't want to have a file descriptor hanging
around until the Gang is garbage collected.